Group misc functions into misc.php on the idp and sp Add Federation Terminaison metadata on the idp and sp Add view off federation on the sp Add cancel federation button on the sp and idp Defederation is not working yet

+++ php/Attic/examples/sample-idp/setup.php
@@ -222,6 +222,16 @@
 
 		print "OK";
 
+		print "<br>Insert user 'test' into 'users' : ";
+		
+		$query = "INSERT INTO users(user_id, username, password, created) ";
+		$query .= "VALUES (nextval('user_id_seq'), 'test', 'test', NOW())";
+		
+		$res =& $db->query($query);
+		if (DB::isError($res)) 
+		  die($res->getMessage());
+		print "OK";
+
 		print "<br>Create table 'nameidentifiers' : ";
 
 		$query = "DROP TABLE nameidentifiers CASCADE";
